python/django/requirements.txt
intersystems-irispython~=3.2.0
django~=4.0.2
djangorestframework~=4.0.2
django-iris~=0.1.5
gunicorn~=20.1.0